requirements.txt
bottle==0.12.9
docopt==0.6.2
dropbox==6.9.0
hkdf==0.0.3
keyring==9.3.1
pickleDB==0.6.2
pycrypto==2.6.1
requests==2.11.1
simplejson==3.8.2
six==1.10.0
typing==3.5.2.2
urllib3==1.18
yagmail==0.4.127
